Isolate error handling of S3 download failure
By pulling the download step into a different try / catch block we can
add logic later to more effectively handle the post-processing cleanup
steps, such as moving the file from `unprocessed` to `processed` in S3.

An earlier implementation of this commit also treated S3 download errors
differently / as a recoverable issue, but after further reflection this
isn't always the right approach (e.g. if the error is a 404) and so I'd
like to think about that improvement at a later date.

Issue #678 Move processed CSV uploads from /unprocessed as part of the processing task