Clean nock mocks between tests
Without this, if a given test sets up a mock via nock that doesn't get
used, that mock will persist to future tests.  This could happen in the
event of an error, for instance.

We don't want our tests to ever have lingering effects!

The auth mocks that we have use nock, so we need to move the setup for
those to beforeEach / afterEach instead of happening just once per
thread.