Move processed CSV uploads as part of processing task
When a bulk CSV upload is completed we want to move it to a permanent
location.  This should happen regardless of whether the upload succeeded
or failed, since we could potentially want the user to be able to access
the file in either situation.

Issue #678 Move processed CSV uploads from /unprocessed as part of the processing task