Add prettier
ESLint is deprecating support for formatting-focused lint rules,
suggesting that projects rely on prettiers instead [1].

This commit adds prettier, though the invocation will be separated into
another commit to make it easier to resolve merge conflicts that might
come up between review and merge.

We decided that single quotes was something we collectively wanted to
keep, so that is the one setting we're overriding.

[1] https://eslint.org/blog/2023/10/deprecating-formatting-rules/