Add support for ignoring formatting commits
We recently rewrote almost every line by invoking prettier to format the
project. This messes with git blame history, but git provides a handy
tool for ignoring that type of commit if the hash is tracked.

This commit adds the hash to a tracking file and provides some
instructions on how to utilize it.