Fix race condition in test
We want this integration test to ensure that the inserted proposal data
properly aligns with the expected base field, but the order in which the
"application form fields" are entered is non deterministic.

This means we need to look up the form fields for the expected base
fields, and ensure that the correct form field is associated with the
proposal data.

Before this fix the tests would usually pass, but in some cases the
insert order might switch, resulting in unpredictable test failures.