Update the file size when processing
We want to actually populate the bulk upload file size as part of the
processing step.  This is actually the soonest we are able to know the
size of the bulk upload, since the presigned post process is separate
from bulk upload registration.

Issue #697 Report BulkUpload file size via API