Change default port to 3001
Prior to this change, the server’s default port (when not configured in
the env vars) was 3000. This happens to conflict with the default port
of the front end — which is set by the `create-react-app` machinery.

While this conflict can be avoided with individual configurations, it
makes sense that our stack is internally compatible out of the box.

Issue #740 Change default port to 3001