Publish to npm for new sdk versions
If the API or the sdk templates increment their version we want to
publish a copy of the package to npm.

The GitHub action we are using here (npm-publish) claims that it will
only publish if the package version has changed from the published
version.

Issue #497 Integrate SDK generation into build process