Only publish to NPM if the NPM_TOKEN secret is set
Secrets are not available to PRs that come from a fork / an outside
contributor.  This was causing, for instance, PRs from dependabot to
break as the sdk publish step was failing.

Only running the publish step in cases where the token is set should
prevent the broken build.  The rest of the sdk build steps are still
performed, which should still protect us from PRs that break the sdk
build more generally.

Issue #771 sdk step is broken for dependabot