Remove `production` versions of scripts
We don't use `npm run` in production so we don't need to have two
versions of `start` and `migrate`.  This takes out the versions of the
scripts that require the build step, and instead just uses the `ts-node`
varieties.

Because we're keeping the simpler names it might look like we took out
the `dev` versions, but really we just made it so there is ONLY a dev
version.

Issue #788 Remove dev versions of commands