Use spaces in `.sql` files
Our existing migrations and queries use spaces, and I suppose in
SQL whitespace is used to form readable queries more than semantic
meaning.  For now I'd like to codify that we're using spaces in sql
files.

We can update to tabs at a later date, but right now the repository is
inconsistent in its editorconfig vs the existing formatting and that's
an immediate problem worth fixing.