Add POST `/organizations` endpoint
For the writable form of this type I wanted to borrow from a cool
utility type that we made for the SDK generation project. This lets us
define readonly attributes of a type and generate the writable form
using the `Writable` utility type.

This also moves us towards the pattern of having explicit exports at the
end of the file (rather than peppering `export` as part of const
definitions)

Issue #777 Allow access to a new /organizations endpoint