Remove the `Applicant` entity and endpoints
The concept of applicant never really got used, and we now know that
proposal data may or may not be associated with an applicant in the PDC.
We also know that applicants are ONLY organizations at this stage of the
PDC, and a proposal might actually be associated with more than one.

For these reasons we are removing the concept of an applicant, and
removing the requirement that an applicant be defined in order for
a proposal to be created.

Issue #777 Allow access to a new /organizations endpoint