Remove package version
We're versioning the API specification, not the API implementation.
This code base isn't being released as a package.

For these reasons we decided NOT to duplicate the `version` number in
both the `openapi.json` as well as `package.json`.  Version is not
a required property in `package.json` so the best solution is to remove
it (rather than have something generic like `0.0.0`

Issue #807 Zero out version in package.json