Stop blocking on `runner.promise`
There was a mistake in our implementation of the job queue -- awaiting
on `runner.promise` would actually block until the runner failed.  This
is what the documentation of graphile worker suggested which is why we
followed that pattern, but a more correct approach for our purposes is
to simply catch errors so that if the worker fails we handle the
failure.

This will allow us to continue to set up other initialization steps
after invoking `startJobQueue` (right now it is not an issue since we
don't ever await on `startJobQueue` but if / when that changes, this
would have caused unexpected behavior.