Support database initialization
We're going to be defining a series of functions in PostgreSQL which
will support extracting structured JSON from the database. These
functions will be treated as code, rather than migrations (i.e. we won't
have a separate copy of the function that changes over time, but will
instead update a single function definition file as needs change).

We still want to ensure the functions are loaded into the database
before the application starts, however, which means we need to improve
the application "start" script to force operation resolution BEFORE the
application starts.

There is a little bit of implicit magic being introduced with this
commit, which is that any `.sql` file in the `database/initialization`
directory will be loaded into the database at application start.
I believe this is akin to the work that nodejs does to add
`node_modules` to the class path / it is a simple hook that is part of
database initialization and so the implicitness is more acceptable
than, say, implicit routing.

Issue #821 Change our query approach to use DB-level json functions