Return deep proposal json for proposal queries
This is the big one...

We're changing the way we populate deep proposals, leveraging psql
directly instead of putting that burden on the application. We may find
that there are optimizations to be made in future, but this removes
a large amount of complexity from the application.

As part of this we are also improving the deep object so that it
includes the base field information directly, meaning a client won't
need to load a list of base fields in order to understand the context of
an application form field.

Issue #821 Change our query approach to use DB-level json functions