Remove type guards from bundle loader
We are in complete control of the interaction with our database and the
structure of the data that comes back. For that reason we've decided to
NOT check the structure of the database return value (and instead we can
rely on integration tests to validate that queries are properly formed,
etc).

This cuts down significantly on code, since we will only need type
guards / schema definitions for the writable form of many of our
entities.  It also means that in runtime we're not re-validating a bunch
of data on GET requests.