Use jest for testing proposal response bodies
In some cases we test bodies using jest, and in other cases we test
response bodies using supertest. It turns out `jest` is a bit better for
this in terms of the test output, but using jest will also allow us to
use things like `expectTimestamp` which is occasionally important.