Add OrganizationProposal endpoints
OrganizationProposals represent relationships between, you guessed it,
organizations and proposals. We had explored the idea of having this
entity be a "nested" entity (e.g. /proposals/:id/organizations) but this
wound up feeling a little bit strange since an endpoint ending in
`/organizations` was returning an `OrganizationProposal` instead of an
`Organization`.

Bottom line, we now have the ability to GET and POST
`/organizationProposals`

Issue #799 Create relationship between Organizations and Proposals