Remove `IdParameter` type
The `IdParameter` type was recently added to the project, and would
have represented a new pattern for extracting an entity id from a given
url string.  However, the new OrganizationProposal entity has made it
clear that paths may well have more than one Id parameter, and so rather
than having an implicit `:id` parameter we should actually have an
explicit `:entityId` parameter (where `entity` is whatever the entity
type might be).

This commit takes out the single use of `IdParameter` and also takes the
opportunity to replace the `:id` parameter with `:opportunityId` in that
single case.