Add validation flag to proposal field and enum to basefield
This commit adds the attribute "is_valid" to the
create-proposal_field_values migration. It also adds an enum to the
base-fields table, which will be used for type validation in the
server. Furthermore, it refactors the seed file to follow a new
convention for typing, where types are just strings, rather than json
literals.  New migrations are 0020 and 0021

Fix Issue #760 add phone number validation