Specify swagger-ui-dist version
We're using a tool caleld `swagger-ui-express` to render the swagger ui
for our API automatically, but we never specify the version of swagger
UI to use.  This means it's locked into whatever the released version
was when we first set it up.

By specifying a version we (A) get the latest but also (B) get future
updates via dependabot.

Issue #851 Update swagger-ui-dist via dependabot