Remove nullable reference arrays
We had a series of "array" attributes of our entities which were
considered nullable.

I believe this is from a time when we allowed shallow objects to be
returned.  Now that deep objects are returned, we can remove the
"nullable" attribute since the array should always be populated.

This will also help with our upcoming migration to OpenAPI v 3.1 since
it has implications on the nullable keyword.