Remove database response validation
We had added a bunch of type guards around database responses, but in
reality we fully control the database response and can "safely" cast the
response to the expected type.

This will simplify our code as well as our tests (and to be clear, we do
still have tests that ensure the format is what we expect -- if the
database returns the wrong format then that will be caught by our
integration tests).