Add Users
We now have the concept of a `user` in the schema, which will allow us
to associate other PDC data with a given user.  This will eventually
support use cases like "list all proposals I have created".

Users, for now, are almost invisible to the API.  They are created (or
selected) based on the `subject` in the authentication JWT.  This is
universal, so all endpoint handlers will have access to the `user`
attribute if one exists for a given session.

Issue #710 Model users in the schema