Decouple authorization and authentication
This commit changes authentication so that it is both a universal step
AND the act of checking authentication is not in itself an act of
authorization.

In addition to this, it adds a simple authorization middleware which
verifies that the request is associated with a valid authentication
object.

This will allow us to add more complex authorization rules to endpoints
over time without duplication of authentication logic, and also allows
us to more predictably access authentication data across ALL endpoints,
regardless of whether or not there are authorization protections.

It also adds some missing tests to our endpoints to ensure that
authentication is properly required (in many cases the check was
logically covered, but these additions make it more explicit)

Issue #870 Consider decoupling authorization from authentication middleware