Replace regex with Phone Package
This commit replaces the current phone number validation method for
proposal field values, which is a hardcoded regex, with the phone
package.  It does not take advantage of the phone package's formatting
functionality, but this could be worth looking into in a future commit
if/when we think about formatting inputted data.  Furthermore, it
refactors the seed file to give relevant base fields the type of phone
number.

Issue #874 Add library for phone number validation