Fail eslint on warning
We don't want warnings from our linter -- it makes noise and represents
something that shouldn't be in the code base.

This commit updates our lint command to fail if there are any warnings.
This will also mean CI fails if there are warnings.

In case anybody is curious, this is NOT an option in `eslintrc` since
it's specific to a given run, so the command line flag is the correct
way to signal an aversion to warnings.