Account for potential baseline users in tests
We had some tests that assumed there would be zero users in the database
when testing user creation; that is not necessarily going to be the case
(e.g. we might have a migration create an administrative user).

This commit updates the tests to verify an offset from that baseline
instead of expecting an absolute value.