Create a test user before each test
The test user would occasionally get created anyway if a given test
invoked the user middleware, but we want to be able to rely on the
existence of a test user for a wide variety of tests which might insert
data directly that is associated with the user BEFORE a given call to
the API is made.

That's what setup is for, so that user has been added!

This also exposed a mistake in our middleware tests (the middleware
tests that a user is created, so it makes sense that user not have the
same identifier as the test user!)