Only return authorized bulk uploads
We now have the concept of bulk upload ownership and so we can change
the behavior of our API to only return the bulk uploads associated with
the current user.  Prior to this commit all users could see all other
user's bulk uploads.

Issue #759 Only return bulk uploads that the logged in user is authorized to view