Use ajv-formats for URL and Emails
This commit adds the ajv-formats package that includes definitions for
the uri and email formats. It replaces the current hardcoded regexes
for those formats with helper functions for validating strings against
the ajv-formats values. Note that we are using the uri format from ajv,
which is in fact a full url with protocol, whereas the uri-reference
format is used for urns(which we are not testing against).

Issue #875 Validate URLs and Emails using ajv formats