Add a system user
The system user is not necessarily going to be used for much aside from
allowing our migrations to run for legacy data where "ownership" is
needed.  For instance, if we add a `created_by` column to the
`bulk_uploads` table we will want to be able to populate that field with
a user.

There may end up being additional use cases for the system user, but
this one is pretty clear and so creating the user makes sense before we
start adding those ownership columns.

Issue #485 Model delegated data ownership in the schema