Create Organizations and OrganizationProposals during bulk upload
We want to be able to make associations between proposals and
organizations during bulk upload if the data allows for it. This first
implementation uses hard coded / special "shortCode" values for base
fields.  At some point we might want to create a more natural way to
detect the appropriate organizational data (or to allow a more
user-driven mapping), but for now this will have to do!

Issue #902 Update bulk upload processing to create organizations and link them to proposals