Add createdBy parameter to /bulkUploads
This adds the ability to specify a createdBy filter when loading bulk
uploads.  It also changes the query so that if the user is an
administrator they will see all bulk uploads unless they specify
a filter.

Issue #928 Add ability to filter by creator
Issue #927 Allow administrative users to view all data regardless of ownership