Add `scope` to BaseField
The scope of a base field signals the type of entity it describes.  We
are starting with just two scopes:

1. Proposal
2. Primary Organization

Proposal base fields would describe a proposal, whereas a primary
organization field would describe the primary organization associated
with a proposal.  This will be useful when attempting to use proposal
data to describe organizations.

Issue #917 Add the concept of "Organization base fields"