Support `me` value for `createdBy` parameters
We want users to be able to specify that they want to retrieve the data
that they created without necessarily knowing their internal user ID.
This change updates the extraction function so that it also takes an
optional auth context.  In the event the auth context is provided, "me"
becomes a valid submission and it is treated as though the user passed
their id.

Issue #960 Create a GET /bulkUploads?mine parameter