Evaluate proposal bundle in upload test
We now have a way to load deep proposals which means we can add a little
bit of simplification to our bulk upload test.  Prior to this we were
loading each individual layer of depth and testing it.

Another reason we were doing this piece by piece approach was in order
to be able to predict various IDs -- this was a lot of added complexity
for not much value (we don't actually care about ID sequence, and so
testing for that isn't necessary)