Fix auth signature test
The previous implementation of this test was overcomplicated (we don't
need to spin up an entire mock server just to generate an improperly
signed token).  It also would occasionally fail because the mock server
used a static set of credentials.

With this change we are now able to generate a signed token directly,
using an alternative private key.  This direction was suggested by the
mock-jwks maintainer [1].

Issue #1009 Our new auth test sometimes fails

[1] https://github.com/levino/mock-jwks/issues/218#issuecomment-2112387014