Update URL to url in field_type enum
This commit remedies a typo from the field_type enum, which is used to
validate the type of a basefield. The value of 'url' was originally
capitalized as 'URL', but this goes against our convention. It adds a
new migration, and updates the seed file accordingly

Issue #989 Update URL enum value to url