Create base_field_localizations table
This commit adds the creation of the base_field_localizations table to
the migration commands, which will be used to store the localizations of
each base field moving forward. The base_field_id field is to be used as
a foreign key on the base_fields table, wherein a base_field may have
multiple localizations. There is a unique constraint on base_field_id
and language to ensure that for any given language, there is only
allowed to be one localization.

Issue #1110 Internationalize Base Fields