Add `DataProvider` entity and routes
A data provider represents an organization that collects and processes
data in the philanthropy landscape and might integrate with PDC to
provide or access data.  Similar to funder it is not described by PDC
field values but it may be the source of a given field value.

Issue #1094 Implement initial provenance support