WIP: add `/organizationDetail/{organizationId}`
The idea is a new endpoint takes an organization ID, finds all the
values associated with that organization, and returns the best values.

At the moment this is not exactly right but it is an outline. It needs
to filter proposal field values by organization, group by base fields,
have an actual function that purifies the gold data, and more.

In order to think best about the ways to do a thing, it helps to try
to do it.

Issue #1087