Add an `extractGold` function stub
This is where I suppose gold extraction will take place. Along the
gold refining analogy lines, we won't end up with purified gold, but
we have a chance at extracting some. Perhaps we are not even to the
point of extracting, we're merely mining. But that has other
connotations in the tech space. We could also abandon the gold analogy
altogether.

The `OrganizationDetail` object is not complete as of this commit. It
still needs the BaseField associated with each value.

Issue #1087