Use GitHub actions to install Java 21
In the SDK workflow, we were previously depending on the version of Java
provided by the runner. Instead, explicitly install the version of Java
we want.

The runner currently provides Java 11 by default[1]. That is older than
the version we claim to require on the SDK[2]. Java 21 is the next LTS,
and the version of Gradle used by the SDK now supports Java 21.

Fixing the version should also help protect us from changes to the
runner image, including if we upgrade to a newer release.

The setup-java action requires a distribution name[3]. Eclipse Temurin
is the new name for AdoptOpenJDK[4].

[1] https://github.com/actions/runner-images/blob/ubuntu22/20240721.1/images/ubuntu/Ubuntu2204-Readme.md
[2] https://github.com/PhilanthropyDataCommons/sdk/blob/49a3de58688b1fe0c7e0d9907f697b77fa0e4157/README.md#prequisites
[3] https://github.com/actions/setup-java?tab=readme-ov-file#supported-distributions
[4] https://blog.adoptopenjdk.net/2021/08/goodbye-adoptopenjdk-hello-adoptium/