Rename fields: anticipate fine-grained permissions
Include the word "visible" in the keys returned for gold data. This is
because we may or may not want to expose a subset of proposal data to
unauthorized parties. At the moment, I assume we will not expose all
proposal field values to all parties once fine-grained permissions are
present. In that case, the system should not say "all" but say
"visible."

If and when different users get different gold data, as would happen
naturally with fine-grained permissions, the set of all visible-to-me
field values will become even more useful (not solely for developers).

Issue #1087