Add base field localizations
This commit adds the concept of base field localizations to the PDC API,
which serve as translations for base field labels and descriptions
across different languages. Base field localizations are unique keyed by
their language and the base field id they are related to, such that a
base field can be related to any number of localizations so long as they
are all unique languages.

Base field localization creation and updating is handled by the nested route
'PUT basefields/{basefieldid}/{language}' using an upsert approach, which is
a new database functionality for the repository, hence the new
'CreateOrUpdate' folder in the operations folder. If an existing basefield
is referenced, it will be updated, else it will be created.

Base fields are now returned with an additional field, 'localizations', which
is a join on the base fields and base_field localizations table, and take the form
of a lookup table relating language to the localization.

Issue #1110 Internationalize Base Fields