Associate Source with ProposalVersion
Proposal versions will have a source association.  For now we don't have
a fully baked mechanism to associate accounts / API actions with
specific sources in a controlled way, so we will just associate all data
with the system source at this stage.  A related issue has been created
to capture the need to improve this part of the system [1].

Until this is thought about and implemented we're not adding sourceId to
the POST of proposal version or to bulk uploads; these two temporary
decisions may eventually change, at which point `sourceId` would no
longer be considered readonly from an external perspective.

Issue #1094 Implement initial provenance support

[1] https://github.com/PhilanthropyDataCommons/service/issues/1146