Use supertest agents sparingly
There is potential (mis)use of shared state across tests when using
supertest agents. Without this change, it can be easy to set, for
example, auth headers on the agent rather than the request and those
headers unintentionally affect tests later in the same file. This
change avoids re-use of state by simply avoiding the use of agent.

In future, there may be some situations that need to use shared state
across tests. In that case, a supertest agent is a fine tool for the
job. See the example at https://www.npmjs.com/package/supertest where
`const agent` is used.

Fixes #1171