Improve Writable utility type
Our `Writable` objects did not support optional parameters, and instead
converted them into required parameters where `undefined` was a valid
type.  This may sound similar, but in practice TypeScript will require
the parameter to be defined (with an explicit `undefined` value) as
opposed to simply resolving to undefined by not existing.

This change to the Writable type means that any optional parameter is
properly mapped to an optional parameter in the resulting type.

You may notice that the mapped type now has `| undefined` as a valid
type, but this is OK since technically any optional parameter DOES allow
`undefined` to be a type.