Add `sourceId` to User
The source ID attribute allows a user to be registered as a valid source
of new data in the PDC. This value will be used to provide provenance
when new field values are added to the system (e.g. new proposal
versions).

For now the value is not being used yet, and in fact there is no way to
set it since there is no endpoint to allow for user creation / editing.

Issue #1154 Support sourceId association with accounts