Add `sourceId` parameter to `POST /proposalVersions`
When we initially implemented provenance we had not yet decided the best
way to actually specify provenance.  This PR reflects the followup
discussion, which is that clients should specify the source of the data
they are uploading.  For now all users with write access are trusted,
and so it is OK to not have guard rails / permission checks to ensure
a user is "allowed" to specify a given source.  That augmentation can be
added later.

Issue #1192 Accept sourceId as a parameter for data insertion API operations