Add structure to `NotFoundError`
By encoding the entity type and identifying information in the
NotFoundError we are able to more effectively translate it into other
error types (e.g. if the search is in the context of a relationship on
POST).

We also just have more value to clients by providing semantically
accessible data.