Fix a test
The test should be testing the .png but in fact there are TWO pieces of
invalid data in the request.  This actually exposes a limitation of our
test suite which is that we're not checking to ensure WHY a given status
code is being returned.  In this case it's possible the 400 was from
another issue.