Organize operations by model
We were organizing database operations by type, not by model.  This was
inconsistent with the way we were organizing queries.  We decided that
organizing by model made more sense, as models are generally going to be
updated together and it is easier to find files that are grouped by
model / there is less scrolling now when making updates.

Issue #1185 Organize DB operations by model at the first level