Avoid union type
By default ajv does not support union types. While we could allow them,
there is reason it is not enabled by default as union types can make
validation take longer in complex entities.

I opted to rewrite this schema in a way that is a little less DRY (we
define the `createdBy` object structure twice), but avoids the union
type.

An alternative approach would be to handle the special `me` case outside
of the schema entirely but I do like the idea of having our schema
defined in a way that truly captures the potential shape of input in one
place.

Issue #1188 Warning when running tests with --no-silent