Add universal `drop_function` sql utility
We use functions to transform rows of data into a consistent json format
when returning values.  We have opted to have a series of functions that
are created as part of server start (as opposed to migrations) to allow
us to flexibly edit the contents of the transformation over time without
having a ton of redundant migrations for minor changes.

This function will allow us to reliably remove past versions of these
functions even in the event that the function signature changes.

Issue #1180 Add universal "drop function" for our initialization queries