Remove use of AuthenticatedRequest typecast for handlers
We were unsafely assigning a type to `req` values in handlers.  Express
does not make any commitments to ensuring that a `req` passed to
a handler is anything more than a `Request`.  If a given handler relies
on attributes of `AuthenticatedRequest` then it would need to explicitly
guard against that.

Issue #1210 Don't use AuthenticatedRequest as the req type