Remove placeholder cases
We had quite a few queries that had cases based on placeholder default
values (e.g. checking if an ID was 0 where "0" was considered
a placeholder non-value).

This replaces that to check for NULL instead, which is cleaner and more
intuitive.

The reason for the complexity before was because I was under the false
impression that `NULL` could not be passed to the query builder!

The only case where I kept a check for a non-null value was the case of
search where an empty string should be considered the same as NULL (e.g.
if user passes an empty value).

Issue #1189 Correct queries that assumed null could not be passed