Update `authenticationId` to `keycloakUserId`
When we first implemented the `User` entity we kept things abstract just
in an interest to avoid coding ourselves down a path when the system was
still in flux.  At this point as a project we're more formally coupled
with Keycloak and it seems appropriate to lean into that relationship.

Given this, we can be more specific that the `authenticationId` is
actually a `keycloakUserId`.  This also means we can lock in that value
as a GUID instead of an arbitrary string.

I did consider using the `username` attribute of KeyCloak, but unlike
the KeyCloak id that can be changed on the keycloak side of things
(which would result in authentication being disassociated with the PDC
authorization state, and that would be a problem).  We may want to store
the keycloak username as a cached attribute and provide the ability to
look things up in terms of the username down the line, but as a primary
key the GUID is a better choice.

We reference the `authenticationId` in other places (e.g. as a filter),
and so this change impacts more endpoints than just the `users`
endpoint.

This improvement / clarity will be relevant as we move towards using
natural keys for our users, and eventually replacing the `id` field with
this field.

Issue #1155 Update users table to use natural key