Create NewType for KeycloakUserId
The library `newtype-ts` allows us to explicitly detach a custom version
of a built in type (e.g. a string) when making type definitions.  In the
case of KeycloakUserId we know that the type is actually a string of
a specific pattern and we don't want to allow any part of our code to
pass in a string that does not fit the pattern.

With this safety comes some additional complexity (we have to explicitly
convert a given `KeycloakUserId` to a `string` before passing it to
something that expects a string), but so it goes.