Show if we're running in production mode
The `NODE_ENV` environment variable should be set to the value
`production` when running a production deployment[1]. The effects that
can have vary from library to library; we're not basing any behavior on
its value, but some of our dependencies might.

In particular, Express improves its performance when
`NODE_ENV=production`[2].

These effects can be subtle, however, so log at startup if we're in
production mode so that we can easily make sure we're deploying
correctly.

This was inspired by my testing a DigitalOcean App deployment of the
service, and wanting to verify that `NODE_ENV=production` was
automatically added without specifically including it in the configured
environment variables; I found that it is automatically included.

[1] https://nodejs.org/en/learn/getting-started/nodejs-the-difference-between-development-and-production
[2] https://expressjs.com/en/advanced/best-practice-performance.html#set-node_env-to-production

Issue PhilanthropyDataCommons/deploy#118 Migrate to DigitalOcean Apps