Rename auth context query parameters
The query parameters were not named in a way that made it clear that
they were communicating the authentication contexted in which a query
was being run (as opposed to, say, a filter value on an entity
attribute).

The new names are a bit more verbose, but are more explicit in their
purpose.