Rename `Organization` to `Changemaker`
The term `Organization` made sense at the time we used it (as
a replacement for `Applicant`), however it is now overly broad and
confusing because we have other types of organization in the PDC (e.g.
funders and data providers).

Changing the term to `Changemaker` allows us to be more specific.  We
had explored the concept of using `Applicant` or `Grantseeker` instead,
but realized not all of the entities stored in the table are necessarily
seeking grants (for instance, they might be partnering changemakers for
a given proposal).

`Changemaker` is not an industry standard term but it does exist in
other parts of the PDC.

We are not changing the base field attributes right now, which does
create a bit of an awkward reality that "changemaker" could
theoretically mean an individual, but the core fields used to map a bulk
upload to a given changemaker is `organization` scoped.  It is possible
this will ... change... in future

Issue #1117 Rename Organization to something more appropriate (semantically close to Changemaker)