Create BaseFieldsCopyTask
This commit creates the new entity 'BaseFieldsCopyTask', along with the
necessary database operations to access and update them. BaseFieldsCopyTasks,
like BulkUploadTasks, are intended to serve as user-accessible entities that
log the status of a basefield copy job between a local and
remote pdc instance. In order to use nock, our current http mocking
framework, for testing, we needed to use an alternative to the
default node fetch package, as that is incompatible with nock.
We decided in the PR to add the node-fetch package, which is
compatible with nock. Notably we added it at an older version,
which per the node-fetch documentation is necessary for commonJs
compatiblity (https://github.com/node-fetch/node-fetch#commonjs)