Pull some shared test data out of gold data tests
This change takes some of the common test data creation out of
individual gold data tests and into a shared `beforeEach` function.
This should not impact test functionality except that each test is more
compact, it requires the reader to reference to variables declared a
little further out from the test, and it takes some of the test setup
that calls production functions out of code coverage, making the test
coverage report more honest.

Issue #1257