Disable old deploys and remove Docker build action
Now that we are using DigitalOcean App Platform to host the API, we no
longer need to trigger the old deploy system. Remove the step that
notifies the PhilanthropyDataCommons/deploy GitHub repo, which also
addresses the problem of needing to periodically rotate the time-limited
personal access token used to do so.

With the old deploy system disabled, the Docker images are no longer
used; they were only ever for us, and we don't know of any users who
relied on the published images. Remove the action that builds the Docker
image and pushes it to Docker Hub.

Issue PhilanthropyDataCommons/deploy#118 Migrate to DigitalOcean Apps
Issue #622 Change deploy model