Pull more shared test data out of gold data tests
This change continues the pattern of extracting some common test setup
into a jest `describe` block and a shared `beforeEach` function. It
does not pull all the setup out and share the entire fixture because
each test tests something different. The different proposals that are
set up for each test should stay within that test. Unfortunately, this
leaves some dishonesty in code coverage when `create...` helpers are
called: those will get undue "coverage" when they are not actually
tested. In some sense, they are "covered", e.g. exercised or called,
but this is far from asserting correct behavior of those helpers.
Separate tests actually cover the helpers with assertions. Perhaps
there is another way to clarify this to the coverage reporter besides
`describe` blocks.

Issue #1257