Add documentation for "gold" changemaker data
Without this commit, one would be left to read the source code to try
to make sense of the values coming out of `GET /changemakers`. While
source code may be well-commented, it does not capture the high-level
intent that drove people to write it. This commit adds documentation
with some details about the `/changemakers` endpoint and special
attention to the method it uses to retrieve field values.

Issue #1274