Use consistent SQL capitalization
The capitalization.types SQLFluff rule requires only that we be
consistent within a file about how we capitalize data types: all
uppercase or all lowercase.

However, in the case of `CREATE FUNCTION..RETURNS`, it seems to consider
that to be its own scope, separate from the function body.  Despite
this, I think the rule is overall worthwhile enough to accept this edge
case.

Use uppercase data types in our queries, and lowercase data types in our
function definitions.

https://docs.sqlfluff.com/en/stable/reference/rules.html#rule-capitalisation.types